requirements.txt
streamlit
streamlit-webrtc
ultralytics
torch
torchvision
Pillow
opencv-python-headless
matplotlib
numpy
